frontend: refactor: separate impurity package
