+++ frontend/src/main/kotlin/compiler/binding/expression/BoundBooleanLiteralExpression.kt
@@ -19,10 +19,10 @@
 package compiler.binding.expression
 
 import compiler.ast.expression.BooleanLiteralExpression
-import compiler.binding.ImpurityVisitor
 import compiler.binding.SideEffectPrediction
 import compiler.binding.context.CTContext
 import compiler.binding.context.ExecutionScopedCTContext
+import compiler.binding.impurity.ImpurityVisitor
 import compiler.binding.type.BoundTypeReference
 import compiler.diagnostic.Diagnosis
 import compiler.diagnostic.NothrowViolationDiagnostic
