+++ frontend/src/main/kotlin/compiler/binding/expression/BoundNullLiteralExpression.kt
@@ -20,10 +20,10 @@ package compiler.binding.expression
 
 import compiler.ast.expression.NullLiteralExpression
 import compiler.ast.type.TypeReference
-import compiler.binding.ImpurityVisitor
 import compiler.binding.SideEffectPrediction
 import compiler.binding.context.CTContext
 import compiler.binding.context.ExecutionScopedCTContext
+import compiler.binding.impurity.ImpurityVisitor
 import compiler.binding.type.BoundTypeReference
 import compiler.diagnostic.Diagnosis
 import compiler.diagnostic.NothrowViolationDiagnostic
