+++ frontend/src/main/kotlin/compiler/binding/expression/BoundReflectExpression.kt
@@ -2,11 +2,11 @@ package compiler.binding.expression
 
 import compiler.ast.expression.AstReflectExpression
 import compiler.ast.type.TypeMutability
-import compiler.binding.ImpurityVisitor
 import compiler.binding.SideEffectPrediction
 import compiler.binding.basetype.BoundBaseType
 import compiler.binding.context.CTContext
 import compiler.binding.context.ExecutionScopedCTContext
+import compiler.binding.impurity.ImpurityVisitor
 import compiler.binding.type.BoundTypeReference
 import compiler.binding.type.RootResolvedTypeReference
 import compiler.binding.type.TypeUseSite
