+++ frontend/src/main/kotlin/compiler/binding/expression/BoundStringLiteralExpression.kt
@@ -2,10 +2,10 @@ package compiler.binding.expression
 
 import compiler.ast.expression.StringLiteralExpression
 import compiler.ast.type.TypeMutability
-import compiler.binding.ImpurityVisitor
 import compiler.binding.SideEffectPrediction
 import compiler.binding.context.CTContext
 import compiler.binding.context.ExecutionScopedCTContext
+import compiler.binding.impurity.ImpurityVisitor
 import compiler.binding.type.BoundTypeReference
 import compiler.diagnostic.Diagnosis
 import compiler.diagnostic.NothrowViolationDiagnostic
