+++ frontend/src/main/kotlin/compiler/binding/impurity/DiagnosingImpurityVisitor.kt
@@ -1,13 +1,9 @@
-package compiler.binding
+package compiler.binding.impurity
 
 import compiler.diagnostic.Diagnosis
 import compiler.diagnostic.PurityViolationDiagnostic
 import compiler.diagnostic.purityViolation
 import compiler.lexer.Span
-
-fun interface ImpurityVisitor {
-    fun visit(impurity: PurityViolationDiagnostic.Impurity)
-}
 
 internal class DiagnosingImpurityVisitor(
     private val diagnosis: Diagnosis,
@@ -16,15 +12,15 @@ internal class DiagnosingImpurityVisitor(
     private val writesBeyondContext = HashSet<Span>()
     private var firstReadSeen = false
 
-    override fun visit(impurity: PurityViolationDiagnostic.Impurity) {
+    override fun visit(impurity: Impurity) {
         when (impurity.kind) {
-            PurityViolationDiagnostic.ActionKind.READ -> {
+            Impurity.ActionKind.READ -> {
                 firstReadSeen = true
                 if (impurity.span !in writesBeyondContext) {
                     diagnosis.purityViolation(impurity, boundaryForReporting)
                 }
             }
-            PurityViolationDiagnostic.ActionKind.MODIFY -> {
+            Impurity.ActionKind.MODIFY -> {
                 check(!firstReadSeen) {
                     "visit writes before reads for this class to work correctly!"
                 }
