+++ frontend/src/main/kotlin/compiler/binding/impurity/ImpurityVisitor.kt
@@ -1,0 +1,5 @@
+package compiler.binding.impurity
+
+fun interface ImpurityVisitor {
+    fun visit(impurity: Impurity)
+}
