+++ frontend/src/main/kotlin/compiler/diagnostic/MutationInConditionDiagnostic.kt
@@ -1,7 +1,9 @@
 package compiler.diagnostic
 
+import compiler.binding.impurity.Impurity
+
 class MutationInConditionDiagnostic(
-    val impurity: PurityViolationDiagnostic.Impurity,
+    val impurity: Impurity,
 ) : Diagnostic(
     Severity.WARNING,
     "This operation is mutating state in a condition.",
